requirements.txt
streamlit==1.56.0
google-generativeai==0.8.6
yt-dlp==2026.3.17
youtube-transcript-api==1.2.4
requests==2.33.1
beautifulsoup4==4.14.3
pandas==3.0.2
tenacity==9.1.4
python-dotenv
supabase==2.13.0
# --- DocGrab Dependencies ---
typer>=0.9.0
markdownify>=0.11.0
lxml>=4.9.0
feedparser==6.0.11
# --- API & Automation ---
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.6
# --- Testing ---
pytest==9.0.3
playwright==1.58.0
pytest-playwright==0.7.2
